zando/database/pyproject.toml
[tool.poetry]
name = "database"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ipykernel = "^6.29.5"
sqlalchemy = "^2.0.40"
cloud-sql-python-connector = "^1.18.0"
numpy = "^2.2.4"
pandas = "^2.2.3"
pydantic = "^2.11.3"
pydantic-settings = "^2.8.1"
pg8000 = "^1.31.2"
pyyaml = "^6.0.2"
fastapi = "^0.115.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gcloud-login = "scripts.gcloud_login:main"
migrate = "db_migrations.cli:main"